requirements.txt
kagglehub==0.2.7
numpy==1.24.4
torch==2.4.0+cu121
torchvision==0.19.0+cu121
pillow==11.1.0
scikit-image==0.25.0
matplotlib==3.10.0
wandb==0.19.11
tqdm==4.67.1